Adding next seminar slides and more examples
git-svn-id: https://svn.code.sf.net/p/cpp-lects-rus/code/trunk@329 39143b06-f351-456c-9f92-4cd32fad6823

+++ other_lects/informatics/code/07-assembly/dbl.c
@@ -1,6 +1,6 @@
-float
-foo (float f1, float f2) {
-  float ftmp;
+double
+foo (double f1, double f2) {
+  double ftmp;
   ftmp = f1 / 3.0;
   f1 = 1.0 - f1;
   f1 = f2 * f1;
